refactor(ui): single source of truth for category emoji
Dashboard and Transactions each carried their own CATEGORY_EMOJI map and they
disagreed: Alimentação was a burger in one and a plate in the other, Transporte
a train against a car, Saúde a pill against a stethoscope, and Dashboard had no
entry for Outros at all. An icon that changes between screens stops identifying
the category, which is the only reason it exists.

The map now lives next to the category lists it mirrors, and the type-based
fallback keeps historical categories rendering.
