+++ frontend/src/lib/categories.js
@@ -33,3 +33,29 @@ export function reconcileCategory(type, current) {
   const list = categoriesFor(type);
   return list.includes(current) ? current : list[0];
 }
+
+// Emoji por categoria. Vive aqui, junto das listas que espelha, porque existiam
+// DOIS mapas divergentes (Dashboard e Relatórios): Alimentação era 🍔 num e 🍽️
+// no outro, Transporte 🚇 contra 🚗, Saúde 💊 contra 🩺. Ícone que muda de tela
+// para tela deixa de identificar a categoria, que é a única função dele.
+const CATEGORY_EMOJI = {
+  "Alimentação": "🍽️",
+  "Moradia": "🏠",
+  "Transporte": "🚗",
+  "Saúde": "🩺",
+  "Educação": "📚",
+  "Lazer": "🎬",
+  "Compras": "🛍️",
+  "Contas & Serviços": "🧾",
+  "Salário": "💼",
+  "Freelance/Extra": "✨",
+  "Investimentos": "📈",
+  "Reembolso": "↩️",
+  "Presente": "🎁",
+  "Outros": "🏷️",
+};
+
+/** Emoji da categoria; cai no genérico por tipo quando a categoria é desconhecida. */
+export function emojiForCategory(category, type) {
+  return CATEGORY_EMOJI[category] ?? (type === "INCOME" ? "🪙" : "💸");
+}
