+++ frontend/src/lib/categories.test.js
@@ -1,6 +1,7 @@
 import { describe, it, expect } from "vitest";
 import {
   categoriesFor,
+  emojiForCategory,
   reconcileCategory,
   EXPENSE_CATEGORIES,
   INCOME_CATEGORIES,
@@ -22,5 +23,19 @@ describe("reconcileCategory", () => {
   it("falls back to the first of the new list when invalid", () => {
     // "Lazer" não existe em receitas → cai na 1ª de receita (Salário)
     expect(reconcileCategory("INCOME", "Lazer")).toBe(INCOME_CATEGORIES[0]);
+  });
+});
+
+describe("emojiForCategory", () => {
+  it("dá um emoji para toda categoria conhecida", () => {
+    for (const c of [...EXPENSE_CATEGORIES, ...INCOME_CATEGORIES]) {
+      expect(emojiForCategory(c, "EXPENSE")).toBeTruthy();
+    }
+  });
+
+  it("cai no genérico por tipo em categoria desconhecida", () => {
+    // Categorias históricas, de antes da lista atual, continuam renderizando.
+    expect(emojiForCategory("Categoria Histórica", "INCOME")).toBe("🪙");
+    expect(emojiForCategory("Categoria Histórica", "EXPENSE")).toBe("💸");
   });
 });
